Tests, bugfixing, reactor flag
Summary:
Filled out a basic test suite which checks if temporary spider class creation produces the correct number of custom settings or rules. Used this test suite to fix several subtle bugs. Also added a flag in SpiderWrangler which allows for crochet to be disabled. To support this, added a wrapper to the start_crawl method (now _start_crawl) which aborts if the reactor was not explicitly started. Attempting to setup crochet multiple times has no effect, so using multiple Wranglers with reactor=True should be safe.

Additions:
+ Added basic tests
* Primarily test if class variable editing is functioning properly
* More tests to be added ASAP
+ reactor flag for SpiderWrangler
* Allows client to determine if the Wrangler initializes a crochet-based Reactor thread on initialization
** This allows for the Wrangler to be used as a metafactory producing temporary spider classes without starting up a costly thread, if the client has another twisted reactor already in place
* Wranglers initialized with reactor=False CANNOT CRAWL. It is sometimes safe, but not guaranteed to be safe
** Initializing multiple Wranglers with reactor=True is safe
+ Fixed error handling-based polymorphism in several methods and utilities

TODO:
* Decide on a license
* Write more tests
* Write a readme
* Write setup.py