Separation of Wrangler and Class Factory concerns
Summary:
I was generally dissatisfied with the conflation of running crawls synchronously and generating new spider classes. The reactor flag from the past commit was an attempt to deal with this conflation by letting clients configure a Wrangler for either crawling or spider production, but this turned out to cause other problems and was ultimately not dealing with the core conflation issue. Spider generation behavior is now encapsulated within SpiderClsFactory and its subclasses. Wranglers are still closely coupled to their associated factories (though not vice-versa), to support configuration of some factory attributes through the Wrangler, but Wranglers can now be configured to either generate temporary spider classes or simply use the provided template using the gen_spiders flag.

Additions:
+ Spider generation behavior now encapsulated within SpiderClsFactory and its subclasses
* SpiderClsFactory - all generic spiders and subclasses except CrawlSpiders
* CrawlSpiderClsFactory - CrawlSpiders and subclasses
+ Wranglers now purely take care of setting up the crochet Reactor and running crawls
+ Edited tests to now test class factories, more tests still on the way
+ rename wrangler_utils.py to factory_utils.py
+ Migrated the function which creates custom_settings dictionaries from factory_utils to a class method within SpiderClsFactory

TODO:
* Better error handling
* Decide on a license
* Write more tests
* Write a readme
* Write setup.py